particles move according to initial velocity

+++ 2024/2024-06-04-star-formation-simulator/rust-simulator/src/types.rs
@@ -1,22 +1,22 @@
 use serde::{Serialize, Deserialize};
 
-#[derive(Debug, Clone, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Vector {
     pub x: f64,
     pub y: f64,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Point {
     pub x: i32,
     pub y: i32,
 }
 
-#[derive(Debug, Serialize, Clone)]
+#[derive(Debug, Serialize)]
 pub struct Particle {
     pub id: usize,
-    pub x: f64,
-    pub y: f64,
+    pub x: u32,
+    pub y: u32,
     pub v: Vector,
     pub mass: f64,
 }
